fix: resolve CodeQL circular import and NoneType raise alerts
Move ExecutionContext from executor.py to execution.py to break the
circular import between interfaces.py and executor.py detected by CodeQL.
Fix ThreadingTimeoutHandler to raise directly from exception_container
instead of via a nullable variable (CodeQL: illegal NoneType raise).
